new changes

+++ project1-backend/src/main/java/com/revature/controllers/UserController.java
@@ -2,6 +2,7 @@ package com.revature.controllers;
 
 import com.revature.models.User;
 import com.revature.services.UserService;
+import jakarta.servlet.http.HttpSession;
 import org.apache.coyote.Response;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
@@ -9,7 +10,8 @@ import org.springframework.web.bind.annotation.*;
 
 @RestController
 @RequestMapping("/users")
-@CrossOrigin(origins = "*")
+@CrossOrigin(origins = "http://localhost:3001", allowCredentials = "true")
+
 public class UserController {
 
     private UserService us;
@@ -35,19 +37,30 @@ public class UserController {
     }
 
     @GetMapping
-    public ResponseEntity<Object> getAllUsers(){
+    public ResponseEntity<Object> getAllUsers(HttpSession session){
+        String role = (String) session.getAttribute("role");
         try{
-            return ResponseEntity.ok(us.getAllUsers());
+            if(role!=null && role.equals("manager")) {
+                return ResponseEntity.ok(us.getAllUsers());
+            }else{
+                return ResponseEntity.status(401).body("You do not have authorization to perform this action.");
+            }
         } catch(IllegalArgumentException e){
             return ResponseEntity.status(400).body(e.getMessage());
         }
     }
 
     @DeleteMapping("/{userId}")
-    public ResponseEntity<Object> deleteUserById(@PathVariable int userId){
+    public ResponseEntity<Object> deleteUserById(@PathVariable int userId, HttpSession session){
+        String role = (String) session.getAttribute("role");
         try{
-            us.deleteUserById(userId);
-            return ResponseEntity.status(200).body("User with id " + userId + " has been deleted");
+            if(role!= null && role.equals("manager")){
+                us.deleteUserById(userId);
+                return ResponseEntity.status(200).body("User with id " + userId + " has been deleted");
+            }else{
+                return ResponseEntity.status(401).body("You do not have authorization to perform this action.");
+            }
+
         } catch (IllegalArgumentException e){
             return ResponseEntity.status(400).body(e.getMessage());
         }
@@ -55,9 +68,15 @@ public class UserController {
     }
 
     @PatchMapping("/{userId}")
-    public ResponseEntity<Object> updateUserRole(@RequestBody String newRole, @PathVariable int userId){
+    public ResponseEntity<Object> updateUserRole(@RequestBody String newRole, @PathVariable int userId, HttpSession session){
+        String role = (String) session.getAttribute("role");
         try{
-            return ResponseEntity.status(200).body(us.updateUserRoleById(newRole, userId));
+            if(role!=null && role.equals("manager")){
+                return ResponseEntity.status(200).body(us.updateUserRoleById(newRole, userId));
+            }else{
+                return ResponseEntity.status(401).body("You do not have authorization to perform this action.");
+            }
+
         } catch(IllegalArgumentException e){
             return ResponseEntity.status(400).body(e.getMessage());
         }
